Crate stable_bst

Maps are collections of unique keys with corresponding values, and sets are just unique keys without a corresponding value.

This crate defines the TreeMap and TreeSet types. Their keys must implement Ord.

TreeMaps are ordered.

Examples

use stable_bst::TreeSet;

let mut tree_set = TreeSet::new();

tree_set.insert(2);
tree_set.insert(1);
tree_set.insert(3);

for i in tree_set.iter() {
   println!("{}", i) // prints 1, then 2, then 3
}

Reexports

pub use map::TreeMap;

pub use set::TreeSet;

Modules

map

set