Struct stable_bst::set::TreeSet

pub struct TreeSet<T, C: Compare<T> = Natural<T>> {
    // some fields omitted
}

An implementation of a set on top of the TreeMap container.

Examples

use stable_bst::TreeSet;

let mut set = TreeSet::new();

set.insert(2);
set.insert(1);
set.insert(3);

for i in set.iter() {
   println!("{}", i) // prints 1, then 2, then 3
}

set.remove(&3);

if !set.contains(&3) {
    println!("set does not contain a 3 anymore");
}

A TreeSet can also be used with a custom ordering:

use stable_bst::TreeSet;

struct Troll<'a> {
    name: &'a str,
    level: u32,
}

let mut trolls = TreeSet::with_comparator(|l: &Troll, r: &Troll| l.level.cmp(&r.level));

trolls.insert(Troll { name: "Orgarr", level: 2 });
trolls.insert(Troll { name: "Blargarr", level: 3 });
trolls.insert(Troll { name: "Kron the Smelly One", level: 4 });
trolls.insert(Troll { name: "Wartilda", level: 1 });

println!("You are facing {} trolls!", trolls.len());

// Print the trolls, ordered by level with smallest level first
for x in trolls.iter() {
    println!("level {}: {}!", x.level, x.name);
}

// Kill all trolls
trolls.clear();
assert_eq!(trolls.len(), 0);

Methods

impl<T: Ord> TreeSet<T>

fn new() -> TreeSet<T>

Creates an empty TreeSet ordered according to the natural order of its values.

Examples

use stable_bst::TreeSet;
let mut set: TreeSet<i32> = TreeSet::new();

impl<T, C> TreeSet<T, C> where C: Compare<T>

fn with_comparator(cmp: C) -> TreeSet<T, C>

Creates an empty TreeSet ordered according to the given comparator.

fn comparator(&self) -> &C

Returns the comparator according to which the TreeSet is ordered.

fn iter(&self) -> Iter<T>

Gets a lazy iterator over the values in the set, in ascending order.

Examples

use stable_bst::TreeSet;
let set: TreeSet<i32> = [1, 4, 3, 5, 2].iter().map(|&x| x).collect();

// Will print in ascending order.
for x in set.iter() {
    println!("{}", x);
}

fn rev_iter(&self) -> RevIter<T>

Gets a lazy iterator over the values in the set, in descending order.

Examples

use stable_bst::TreeSet;
let set: TreeSet<i32> = [1, 4, 3, 5, 2].iter().map(|&x| x).collect();

// Will print in descending order.
for x in set.rev_iter() {
    println!("{}", x);
}

fn into_iter(self) -> IntoIter<T>

Creates a consuming iterator, that is, one that moves each value out of the set in ascending order. The set cannot be used after calling this.

Examples

use stable_bst::TreeSet;
let set: TreeSet<i32> = [1, 4, 3, 5, 2].iter().map(|&x| x).collect();

// Not possible with a regular `.iter()`
let v: Vec<i32> = set.into_iter().collect();
assert_eq!(v, vec![1, 2, 3, 4, 5]);

fn lower_bound(&self, v: &T) -> Iter<T>

Gets a lazy iterator pointing to the first value not less than v (greater or equal). If all elements in the set are less than v empty iterator is returned.

Examples

use stable_bst::TreeSet;
let set: TreeSet<i32> = [2, 4, 6, 8].iter().map(|&x| x).collect();

assert_eq!(set.lower_bound(&4).next(), Some(&4));
assert_eq!(set.lower_bound(&5).next(), Some(&6));
assert_eq!(set.lower_bound(&10).next(), None);

fn upper_bound(&self, v: &T) -> Iter<T>

Gets a lazy iterator pointing to the first value greater than v. If all elements in the set are less than or equal to v an empty iterator is returned.

Examples

use stable_bst::TreeSet;
let set: TreeSet<i32> = [2, 4, 6, 8].iter().map(|&x| x).collect();

assert_eq!(set.upper_bound(&4).next(), Some(&6));
assert_eq!(set.upper_bound(&5).next(), Some(&6));
assert_eq!(set.upper_bound(&10).next(), None);

fn difference<'a>(&'a self, other: &'a TreeSet<T, C>) -> Difference<'a, T, C> where C: Eq

Visits the values representing the difference, in ascending order.

Examples

use stable_bst::TreeSet;

let a: TreeSet<i32> = [1, 2, 3].iter().map(|&x| x).collect();
let b: TreeSet<i32> = [3, 4, 5].iter().map(|&x| x).collect();

// Can be seen as `a - b`.
for x in a.difference(&b) {
    println!("{}", x); // Print 1 then 2
}

let diff: TreeSet<i32> = a.difference(&b).map(|&x| x).collect();
assert_eq!(diff, [1, 2].iter().map(|&x| x).collect());

// Note that difference is not symmetric,
// and `b - a` means something else:
let diff: TreeSet<i32> = b.difference(&a).map(|&x| x).collect();
assert_eq!(diff, [4, 5].iter().map(|&x| x).collect());

fn symmetric_difference<'a>(&'a self, other: &'a TreeSet<T, C>) -> SymmetricDifference<'a, T, C> where C: Eq

Visits the values representing the symmetric difference, in ascending order.

Examples

use stable_bst::TreeSet;

let a: TreeSet<i32> = [1, 2, 3].iter().map(|&x| x).collect();
let b: TreeSet<i32> = [3, 4, 5].iter().map(|&x| x).collect();

// Print 1, 2, 4, 5 in ascending order.
for x in a.symmetric_difference(&b) {
    println!("{}", x);
}

let diff1: TreeSet<i32> = a.symmetric_difference(&b).map(|&x| x).collect();
let diff2: TreeSet<i32> = b.symmetric_difference(&a).map(|&x| x).collect();

assert_eq!(diff1, diff2);
assert_eq!(diff1, [1, 2, 4, 5].iter().map(|&x| x).collect());

fn intersection<'a>(&'a self, other: &'a TreeSet<T, C>) -> Intersection<'a, T, C> where C: Eq

Visits the values representing the intersection, in ascending order.

Examples

use stable_bst::TreeSet;

let a: TreeSet<i32> = [1, 2, 3].iter().map(|&x| x).collect();
let b: TreeSet<i32> = [2, 3, 4].iter().map(|&x| x).collect();

// Print 2, 3 in ascending order.
for x in a.intersection(&b) {
    println!("{}", x);
}

let diff: TreeSet<i32> = a.intersection(&b).map(|&x| x).collect();
assert_eq!(diff, [2, 3].iter().map(|&x| x).collect());

fn union<'a>(&'a self, other: &'a TreeSet<T, C>) -> Union<'a, T, C> where C: Eq

Visits the values representing the union, in ascending order.

Examples

use stable_bst::TreeSet;

let a: TreeSet<i32> = [1, 2, 3].iter().map(|&x| x).collect();
let b: TreeSet<i32> = [3, 4, 5].iter().map(|&x| x).collect();

// Print 1, 2, 3, 4, 5 in ascending order.
for x in a.union(&b) {
    println!("{}", x);
}

let diff: TreeSet<i32> = a.union(&b).map(|&x| x).collect();
assert_eq!(diff, [1, 2, 3, 4, 5].iter().map(|&x| x).collect());

fn len(&self) -> usize

Return the number of elements in the set

Examples

use stable_bst::TreeSet;

let mut v = TreeSet::new();
assert_eq!(v.len(), 0);
v.insert(1);
assert_eq!(v.len(), 1);

fn is_empty(&self) -> bool

Returns true if the set contains no elements

Examples

use stable_bst::TreeSet;

let mut v = TreeSet::new();
assert!(v.is_empty());
v.insert(1);
assert!(!v.is_empty());

fn clear(&mut self)

Clears the set, removing all values.

Examples

use stable_bst::TreeSet;

let mut v = TreeSet::new();
v.insert(1);
v.clear();
assert!(v.is_empty());

fn contains<Q: ?Sized>(&self, value: &Q) -> bool where C: Compare<Q, T>

Returns true if the set contains a value.

The value may be any borrowed form of the set's value type, but the ordering on the borrowed form must match the ordering on the value type.

Examples

use stable_bst::TreeSet;

let set: TreeSet<i32> = [1, 2, 3].iter().map(|&x| x).collect();
assert_eq!(set.contains(&1), true);
assert_eq!(set.contains(&4), false);

fn is_disjoint(&self, other: &TreeSet<T, C>) -> bool where C: Eq

Returns true if the set has no elements in common with other. This is equivalent to checking for an empty intersection.

Examples

use stable_bst::TreeSet;

let a: TreeSet<i32> = [1, 2, 3].iter().map(|&x| x).collect();
let mut b: TreeSet<i32> = TreeSet::new();

assert_eq!(a.is_disjoint(&b), true);
b.insert(4);
assert_eq!(a.is_disjoint(&b), true);
b.insert(1);
assert_eq!(a.is_disjoint(&b), false);

fn is_subset(&self, other: &TreeSet<T, C>) -> bool where C: Eq

Returns true if the set is a subset of another.

Examples

use stable_bst::TreeSet;

let sup: TreeSet<i32> = [1, 2, 3].iter().map(|&x| x).collect();
let mut set: TreeSet<i32> = TreeSet::new();

assert_eq!(set.is_subset(&sup), true);
set.insert(2);
assert_eq!(set.is_subset(&sup), true);
set.insert(4);
assert_eq!(set.is_subset(&sup), false);

fn is_superset(&self, other: &TreeSet<T, C>) -> bool where C: Eq

Returns true if the set is a superset of another.

Examples

use stable_bst::TreeSet;

let sub: TreeSet<i32> = [1, 2].iter().map(|&x| x).collect();
let mut set: TreeSet<i32> = TreeSet::new();

assert_eq!(set.is_superset(&sub), false);

set.insert(0);
set.insert(1);
assert_eq!(set.is_superset(&sub), false);

set.insert(2);
assert_eq!(set.is_superset(&sub), true);

fn insert(&mut self, value: T) -> bool

Adds a value to the set. Returns true if the value was not already present in the set.

Examples

use stable_bst::TreeSet;

let mut set = TreeSet::new();

assert_eq!(set.insert(2), true);
assert_eq!(set.insert(2), false);
assert_eq!(set.len(), 1);

fn remove<Q: ?Sized>(&mut self, value: &Q) -> bool where C: Compare<Q, T>

Removes a value from the set. Returns true if the value was present in the set.

The value may be any borrowed form of the set's value type, but the ordering on the borrowed form must match the ordering on the value type.

Examples

use stable_bst::TreeSet;

let mut set = TreeSet::new();

set.insert(2);
assert_eq!(set.remove(&2), true);
assert_eq!(set.remove(&2), false);