Module stable_bst::map

Structs

IntoIter
Lazy forward iterator over a map that consumes the map while iterating

Iter
Lazy forward iterator over a map

IterMut
Lazy forward iterator over a map that allows for the mutation of the values.

Keys
TreeMap keys iterator.

Range

RangeMut

TreeMap
This is implemented as an AA tree, which is a simplified variation of a red-black tree where red (horizontal) nodes can only be added as a right child. The time complexity is the same, and re-balancing operations are more frequent but also cheaper.

Values
TreeMap values iterator.

ValuesMut
TreeMap values iterator.

Enums

Backward

Forward

Traits

Direction